front 2 websockets for game and user status
